resolve jdl circular dependency
