+++ jdl/converters/types.ts
@@ -1,5 +1,4 @@
-export type RelationshipSide = 'left' | 'right';
-export type RelationshipType = 'one-to-one' | 'many-to-one' | 'one-to-many' | 'many-to-many';
+import { RelationshipType, RelationshipSide } from '../basic-types/relationships.js';
 
 export type Field = {
   fieldName: string;
