+++ jdl/jhipster/index.mts
@@ -26,6 +26,7 @@ export { default as serviceDiscoveryTypes } from './service-discovery-types.js';
 export { default as testFrameworkTypes } from './test-framework-types.js';
 export { default as unaryOptions } from './unary-options.js';
 export { default as validations } from './validations.js';
+export * from './validations.js';
 export { default as websocketTypes } from './websocket-types.js';
 export { default as checkAndReturnRelationshipOnValue } from './relationship-on-handler-options.js';
 export { IngressTypes as ingressTypes } from './kubernetes-platform-types.js';
